* commit 'c983c3711028afa7e4d60f9f103f12f9a11df0b3': (112 commits)
  Correction to allow test suite upload when behind a reverse proxy
  Corrected a DB deadlock when deleting a test suite with test cases.
  Correction to allow test suite upload when behind a reverse proxy
  Corrected FK issue
  Temporary solution for v1.2.0 to avoid caching of JS and CSS sources
  Corrected injection for angular controllers and services
  [ITB-194] Display of version text included
  [ITB-112, ITB-153] Prevented basic users of modifying conformance statements
  [ITB-47, ITB-74, ITB-68] Re-implemented the management of systems
  [ITB-113] Basic user can't change endpoints
  [ITB-47, ITB-74, ITB-68] Re-implemented the management of systems
  [ITB-146] Corrected display of conformance testing status for system
  [ITB-140] Included step title in report
  [ITB-104] Server side updates are now synchronous
  [ITB-75] Corrected test session overview report when test case is deleted
  [ITB-145] Removed display of options from wizard
  [ITB-146] Corrected display of conformance testing status for system
  [ITB-75] Test result history is now maintained for deletions and updates
  [ITB-151] System indented display corrected
  [ITB-146] Corrected display of conformance testing status for system
  ...

+++ gitb-engine/src/main/java/com/gitb/engine/SessionManager.java
@@ -31,6 +31,14 @@ public class SessionManager {
 		return instance;
 	}
 
+	public boolean exists(String sessionId) {
+		return contexts.containsKey(sessionId);
+	}
+
+	public boolean notExists(String sessionId) {
+		return !exists(sessionId);
+	}
+
     /**
      * Create a new testcase execution session
      * @param testCaseId
@@ -41,6 +49,8 @@ public class SessionManager {
 		String sessionId = UUID.randomUUID().toString();
 		//Load the tdl:TestCase definition
         TestCase testCase = TestCaseManager.getTestCaseDescription(testCaseId);
+        // Ensure we replace the text ID with the internal fully unique ID
+		testCase.setId(testCaseId);
         //Create the test case context
 		TestCaseContext testCaseContext = new TestCaseContext(testCase, sessionId);
         //Put the context into the map
@@ -65,7 +75,9 @@ public class SessionManager {
 	public void endSession(String sessionId) {
 		TestCaseContext testCaseContext = contexts.remove(sessionId);
 
-		testCaseContext.destroy();
+		if (testCaseContext != null) {
+			testCaseContext.destroy();
+		}
 	}
 
 	public void destroy() {
